changed file structure and resoldered encoder to interrupt pins
