+++ Release/Bot_builder.cpp.d
@@ -1,0 +1,22 @@
+C:\Users\Spencer\AppData\Local\Temp\VMBuilds\Bot_builder\mega_atmega2560\Release\Bot_builder.cpp.o: \
+ C:\Users\Spencer\AppData\Local\Temp\VMBuilds\Bot_builder\mega_atmega2560\Release\Bot_builder.cpp \
+ C:\Program\ Files\ (x86)\Arduino\hardware\arduino\avr\cores\arduino/arduino.h \
+ C:\Program\ Files\ (x86)\Arduino\hardware\arduino\avr\cores\arduino/binary.h \
+ C:\Program\ Files\ (x86)\Arduino\hardware\arduino\avr\cores\arduino/WCharacter.h \
+ C:\Program\ Files\ (x86)\Arduino\hardware\arduino\avr\cores\arduino/WString.h \
+ C:\Program\ Files\ (x86)\Arduino\hardware\arduino\avr\cores\arduino/HardwareSerial.h \
+ C:\Program\ Files\ (x86)\Arduino\hardware\arduino\avr\cores\arduino/Stream.h \
+ C:\Program\ Files\ (x86)\Arduino\hardware\arduino\avr\cores\arduino/Print.h \
+ C:\Program\ Files\ (x86)\Arduino\hardware\arduino\avr\cores\arduino/Printable.h \
+ C:\Program\ Files\ (x86)\Arduino\hardware\arduino\avr\cores\arduino/USBAPI.h \
+ C:\Program\ Files\ (x86)\Arduino\hardware\arduino\avr\cores\arduino/Arduino.h \
+ C:\Program\ Files\ (x86)\Arduino\hardware\arduino\avr\variants\mega/pins_arduino.h \
+ C:\Users\Spencer\Documents\Arduino\libraries\Encoder-master/Encoder.h \
+ C:\Program\ Files\ (x86)\Arduino\hardware\arduino\avr\cores\arduino/Arduino.h \
+ C:\Users\Spencer\Documents\Arduino\libraries\Encoder-master/utility/direct_pin_read.h \
+ C:\Users\Spencer\Documents\Arduino\libraries\Encoder-master/utility/interrupt_pins.h \
+ C:\Users\Spencer\AppData\Local\Temp\VMBuilds\Bot_builder\mega_atmega2560\Release\Wires.h \
+ C:\Users\Spencer\AppData\Local\Temp\VMBuilds\Bot_builder\mega_atmega2560\Release\src/lineFollowing/lineFollowing.h \
+ C:\Users\Spencer\AppData\Local\Temp\VMBuilds\Bot_builder\mega_atmega2560\Release\src/lineFollowing/src/LineSensor/LineSensor.h \
+ C:\Users\Spencer\AppData\Local\Temp\VMBuilds\Bot_builder\mega_atmega2560\Release\src/lineFollowing/src/Motors/Motors.h \
+ C:\Users\Spencer\AppData\Local\Temp\VMBuilds\Bot_builder\mega_atmega2560\Release\src/DeadReckoning/DeadReckoner.h
