+++ Release/src/DistanceSensor/disSensor.cpp.d
@@ -1,0 +1,14 @@
+C:\Users\Spencer\AppData\Local\Temp\VMBuilds\Bot_builder\mega_atmega2560\Release\src\DistanceSensor\disSensor.cpp.o: \
+ C:\Users\Spencer\AppData\Local\Temp\VMBuilds\Bot_builder\mega_atmega2560\Release\src\DistanceSensor\disSensor.cpp \
+ C:\Users\Spencer\AppData\Local\Temp\VMBuilds\Bot_builder\mega_atmega2560\Release\src\DistanceSensor\disSensor.h \
+ C:\Program\ Files\ (x86)\Arduino\hardware\arduino\avr\cores\arduino/Arduino.h \
+ C:\Program\ Files\ (x86)\Arduino\hardware\arduino\avr\cores\arduino/binary.h \
+ C:\Program\ Files\ (x86)\Arduino\hardware\arduino\avr\cores\arduino/WCharacter.h \
+ C:\Program\ Files\ (x86)\Arduino\hardware\arduino\avr\cores\arduino/WString.h \
+ C:\Program\ Files\ (x86)\Arduino\hardware\arduino\avr\cores\arduino/HardwareSerial.h \
+ C:\Program\ Files\ (x86)\Arduino\hardware\arduino\avr\cores\arduino/Stream.h \
+ C:\Program\ Files\ (x86)\Arduino\hardware\arduino\avr\cores\arduino/Print.h \
+ C:\Program\ Files\ (x86)\Arduino\hardware\arduino\avr\cores\arduino/Printable.h \
+ C:\Program\ Files\ (x86)\Arduino\hardware\arduino\avr\cores\arduino/USBAPI.h \
+ C:\Program\ Files\ (x86)\Arduino\hardware\arduino\avr\cores\arduino/Arduino.h \
+ C:\Program\ Files\ (x86)\Arduino\hardware\arduino\avr\variants\mega/pins_arduino.h
