+++ Wires.h
@@ -8,13 +8,13 @@
 #define RIGHT_MOTOR_TOP_WIRE 10
 #define RIGHT_MOTOR_BOTTOM_WIRE 11
 
-#define RIGHT_LINE_SENSOR 2
-#define LEFT_LINE_SENSOR 3
+#define RIGHT_LINE_SENSOR 12
+#define LEFT_LINE_SENSOR 13
 
-#define LEFT_MOTOR_BCD_1 50
-#define LEFT_MOTOR_BCD_2 51
-#define RIGHT_MOTOR_BCD_1 52
-#define RIGHT_MOTOR_BCD_2 53
+#define RIGHT_MOTOR_BCD_WHITE_B 18
+#define RIGHT_MOTOR_BCD_YELLOW_A 19
+#define LEFT_MOTOR_BCD_WHITE_B 21
+#define LEFT_MOTOR_BCD_YELLOW_A 20
 
 #define FRONT_LEFT_DISTANCE 0
 #define FRONT_RIGHT_DISTANCE 1
