+++ __vm/.Bot_builder.vsarduino.h
@@ -5,22 +5,18 @@
 			all non-arduino files created by visual micro and all visual studio project or solution files can be freely deleted and are not required to compile a sketch (do not delete your own code!).
 			note: debugger breakpoints are stored in '.sln' or '.asln' files, knowledge of last uploaded breakpoints is stored in the upload.vmps.xml file. Both files are required to continue a previous debug session without needing to compile and upload again
 	
-	Hardware: Arduino Yún, Platform=avr, Package=arduino
+	Hardware: Arduino/Genuino Mega w/ ATmega2560 (Mega 2560), Platform=avr, Package=arduino
 */
 
 #if defined(_VMICRO_INTELLISENSE)
 
 #ifndef _VSARDUINO_H_
 #define _VSARDUINO_H_
-#define __AVR_ATmega32u4__
-#define __AVR_ATmega32U4__
-#define _VMDEBUG 1
+#define __AVR_ATmega2560__
 #define F_CPU 16000000L
 #define ARDUINO 10807
-#define ARDUINO_AVR_YUN
+#define ARDUINO_AVR_MEGA2560
 #define ARDUINO_ARCH_AVR
-#define USB_VID 0x2341
-#define USB_PID 0x8041
 #define __cplusplus 201103L
 #define __AVR__
 #define __inline__
@@ -78,6 +74,6 @@ typedef void *__builtin_va_list;
 #define pgm_read_float(address_short) float()
 #define pgm_read_ptr(address_short)   short()
 
-#include "lineFollowing.ino"
+#include "Bot_builder.ino"
 #endif
 #endif
