+++ src/lineFollowing/lineFollowing.h
@@ -1,5 +1,8 @@
 #ifndef lineFollowing_H
 #define lineFollowing_H
+#include "src/LineSensor/LineSensor.h"
+#include "src/Motors/Motors.h"
+
 void lineFollow(Motors moters,LineSensor leftSide,LineSensor rightSide);
 
 #endif
