--- src/lineFollowing/lineFollowing.ino
@@ -1,75 +1,0 @@
-#include <Encoder.h>
-
-#include "LineSensor.h"
-#include "Motors.h"
-#include "Wires.h"
-#include "lineFollowing.h"
-#include "DeadReckoner.h"
-//#include "Odometry.h"
-
-Encoder knobLeft(LEFT_MOTOR_BCD_YELLOW_A, LEFT_MOTOR_BCD_WHITE_B);
-Encoder knobRight(RIGHT_MOTOR_BCD_YELLOW_A, RIGHT_MOTOR_BCD_WHITE_B);
-
-LineSensor leftSide(LEFT_LINE_SENSOR), rightSide(RIGHT_LINE_SENSOR);
-
-const int mr_en = RIGHT_MOTOR_ENABLE; //Motor 1 PWM Speed Control Pin, Pin 1 on L239
-const int ml_en = LEFT_MOTOR_ENABLE; //Motor 2 PWM Speed Control Pin, Pin 9 on L239
-
-const int mr_da = RIGHT_MOTOR_TOP_WIRE; //Motor 1 Directional Control A, Pin 2 on L239
-const int mr_db = RIGHT_MOTOR_BOTTOM_WIRE; //Motor 1 Directional Control B, Pin 7 on L239
-const int ml_da = LEFT_MOTOR_TOP_WIRE; //Motor 2 Directional Control A, Pin 15 on L239
-const int ml_db = LEFT_MOTOR_BOTTOM_WIRE; //Motor 2 Directional Control B, Pin 10 on L239
-
-long positionLeft  = -999;
-long positionRight = -999;
-//auto timer = millis();
-
-//int i = 0, j = 0, k = 0;
-
-//int left[1000];
-//int right[1000];
-
-Motors motors(mr_en, ml_en, mr_da, mr_db, ml_da, ml_db);
-
-void setup() {
-  Serial.begin(9600);
-  Serial.println("Two Wheels Encoder Test: ");
-  
-}
-void loop() {
-  
- //lineFollow(motors, leftSide,rightSide);
-
-  //Encoder's view while driving:
-  long newLeft, newRight;
-  
-//  left[i] = knobLeft.read();
-//  i++;
-//  right[j] = knobRight.read();
-//  j++;
-  newLeft = knobLeft.read();
-  newRight = knobRight.read();
-  if (newLeft != positionLeft || newRight != positionRight) {
-    Serial.print("Left = ");
-    Serial.print(newLeft);
-    Serial.print(", Right = ");
-    Serial.print(newRight);
-    Serial.println();
-    positionLeft = newLeft;
-    positionRight = newRight;
-  }
-  // if a character is sent from the serial monitor,
-  // reset both back to zero.
-  if (Serial.available()) {
-    Serial.read();
-    Serial.println("Reset both knobs to zero");
-    knobLeft.write(0);
-    knobRight.write(0);
-  }
-  motors.drive();
-// for(; k < i; k++){
-//  Serial.print("left clicks = "); 
-//  Serial.println(left[k]);
-// }
-delay(1000);
-}
